envoy: Use local copies of required BPF header files
Depending on the version of the system headers available when compilation
happens, the version of <linux/bpf.h> available may be insufficient.

There are two possible options to address this:
 1. Refer to the top cilium directory as an external bazel dependency.
 2. Provide a local copy of the header files

Given that these headers are rarely ever going to change, making a local copy
of these header file is chosen for simplicity.

Signed-off-by: Thomas Graf <thomas@cilium.io>