daemon: revisit restore endpoint functionality
Daemon couldn't serve the API immediately as it would restore all
previous endpoints. That was a slow process due the lack of running
time-consuming operations on the background. For this reason, this
commit moves such operations to be executed on the background, leaving
the API in ready state almost immediately.

When restoring an endpoint there are steps that need to be done in this
particular order:
 1 - Parse endpoint for the lxc_config.h;
 2 - Check if the endpoint has its container running;
 3 - Allocate its IP addresses;
 4 - Set default endpoint options;
 5 - Synchronize endpoints' labels with the kv-store;
 6 - Regenerate endpoint.

The necessary steps that needed to be executed before the API is opened
are 1 to 4, which takes relatively less time than synchronizing the
labels with the kv-store or regenerating the endpoint. After step 4 is
executed for all endpoints, the API is opened.

To inform the user of such restore operations, a new endpoint state was
created designated by "restoring" which allows the user to verify which
endpoints are being in restoring state once cilium is started.

With this commit cilium takes around 3 to 4 seconds to have its API
available while restoring 30 endpoints.

Signed-off-by: André Martins <andre@cilium.io>